refactor(services): extract team role helpers into dedicated module
Move get_team_role_map, invalidate_team_role_cache, and extract_team_from_roles functions from discord_helpers.py to a new team_helpers.py module. Add backward-compatibility import in discord_helpers.py to maintain existing import paths. This refactoring improves code organization and supports the platform-agnostic architecture by separating Discord-specific helpers from shared team utilities.
